REFACT: add static
