+++ src/parse/is_valid/is_rt_file.c
@@ -1,0 +1,16 @@
+#include "../../../include/parse.h"
+
+t_is_return	is_rt_file(char *file_name)
+{
+	int		len;
+	char	*file_ext;
+	t_is_return	match;
+
+	match = NO;
+	len = ft_strlen(file_name);
+	file_ext = ft_substr(file_name, len - 3, 3);
+	if (ft_strncmp(file_ext, ".rt", 3) == 0)
+		match = YES;
+	free(file_ext);
+	return (match);
+}
