+++ src/parse/is_valid/parse_elements.c
@@ -1,21 +1,6 @@
 #include "../../../include/parse.h"
 
-t_is_return	is_rt_file(char *file_name)
-{
-	int		len;
-	char	*file_ext;
-	t_is_return	match;
-
-	match = NO;
-	len = ft_strlen(file_name);
-	file_ext = ft_substr(file_name, len - 3, 3);
-	if (ft_strncmp(file_ext, ".rt", 3) == 0)
-		match = YES;
-	free(file_ext);
-	return (match);
-}
-
-t_is_return	is_valid_type(char *type)
+static t_is_return	is_valid_type(char *type)
 {
 	static int	capital_type[3];
 
@@ -36,7 +21,7 @@ t_is_return	is_valid_type(char *type)
 	return (YES);
 }
 
-t_is_return	is_valid_elements(char **elements)
+static t_is_return	is_valid_elements(char **elements)
 {
 	char		*type;
 	t_is_return	ret;
